+++ examples/certilizer-cli.sh
@@ -29,3 +29,12 @@ certilizer --conf-file certilizer.yaml --out-format html --out-file ../stage/exa
 printf "\n\n========================================\n"
 printf "Run command with specified maximum column size and expiry threshold in 180 days:\n"
 certilizer --conf-file certilizer.yaml --out-format html --max-col-size 20 --expiry-threshold-in-days 180 --out-file ../stage/example-output-180days.html
+
+printf "\n\n========================================\n"
+printf "Run command with specified json output format, and output file:\n"
+certilizer --conf-file certilizer.yaml --out-format json --out-file ../stage/example-output.json
+
+printf "\n\n========================================\n"
+printf "Run command with specified yaml output format, and output file:\n"
+certilizer --conf-file certilizer.yaml --out-format yaml --out-file ../stage/example-output.yaml
+