+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add CodeQL scanning workflow and badge
 - Add ssl_verify YAML configuration property #2
+- Add JSON and YAML output formats
 
 ### Changed
 - Certificate OCSP is now optional #4